pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "medical-ocr-extractor"
version = "1.0.0"
description = "A FastAPI service that extracts structured medical data from PDF and image reports using Tesseract OCR"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Medical OCR Team"},
]
keywords = ["medical", "ocr", "api", "fastapi", "tesseract", "pdf", "extraction"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pytesseract>=0.3.13",
    "pillow>=11.0.0",
    "pymupdf>=1.26.0",
    "aiofiles>=24.0.0",
    "pydantic>=2.11.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/medical-ocr-extractor"
Repository = "https://github.com/yourusername/medical-ocr-extractor"
Issues = "https://github.com/yourusername/medical-ocr-extractor/issues"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true